[riscv32-cluster] Enhancement: Refactoring Clock Interface
